backend/requirements_full.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
textblob
scikit-learn
pandas
numpy
PyPDF2
python-docx
spacy
nltk
openpyxl

# PostgreSQL Database
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# New dependencies for Talent Intelligence System
twilio==8.10.0
pyresparser==1.0.6

# Face Recognition for Attendance - FULL VERSION
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.1.0
dlib==19.24.2

# QR Code Generation
qrcode[pil]==7.4.2

# Additional ML dependencies
scipy
matplotlib
seaborn